+++ src/services/Xml.php
@@ -2,18 +2,24 @@
 namespace fostercommerce\shipstationconnect\services;
 
 use fostercommerce\shipstationconnect\Plugin;
+use fostercommerce\shipstationconnect\events\OrderFieldEvent;
 use craft\commerce\Plugin as CommercePlugin;
 use craft\commerce\elements\Order;
 use craft\commerce\models\LineItem;
 use craft\commerce\models\Customer;
 use craft\commerce\models\Address;
 use yii\base\Component;
+use yii\base\Event;
 
 class Xml extends Component
 {
+    const ORDER_FIELD_EVENT = 'orderFieldEvent';
+
     public function shouldInclude($order)
     {
-        return $order->getShippingAddress() && $order->getBillingAddress();
+        $settings = Plugin::getInstance()->settings;
+        $billingSameAsShipping = $settings->billingSameAsShipping;
+        return $order->getShippingAddress() && ($billingSameAsShipping || $order->getBillingAddress());
     }
 
     /**
@@ -54,13 +60,27 @@ class Xml extends Component
                     $settings =  Plugin::getInstance()->settings;
                     $prefix = $settings->orderIdPrefix;
                     return $prefix . $order->id;
-                }
+                },
+                'cdata' => false,
             ],
-            'OrderNumber' => 'number',
+            'OrderNumber' => [
+                'callback' => function ($order) {
+                    $orderFieldEvent = new OrderFieldEvent([
+                        'field' => OrderFieldEvent::FIELD_ORDER_NUMBER,
+                        'order' => $order,
+                        'value' => $order->reference,
+                    ]);
+
+                    Event::trigger(static::class, self::ORDER_FIELD_EVENT, $orderFieldEvent);
+                    return $orderFieldEvent->value;
+                },
+                'cdata' => false,
+            ],
             'OrderStatus' => [
                 'callback' => function ($order) {
                     return $order->getOrderStatus()->handle;
-                }
+                },
+                'cdata' => false,
             ],
             'OrderTotal' => [
                 'callback' => function ($order) {
@@ -118,9 +138,17 @@ class Xml extends Component
      */
     public function shippingMethod(\SimpleXMLElement $order_xml, $order)
     {
-        if ($shippingMethod = $order->getShippingMethod()) {
-            $this->addChildWithCDATA($order_xml, 'ShippingMethod', $shippingMethod->handle);
+        $orderFieldEvent = new OrderFieldEvent([
+            'field' => OrderFieldEvent::FIELD_SHIPPING_METHOD,
+            'order' => $order,
+        ]);
+        Event::trigger(static::class, self::ORDER_FIELD_EVENT, $orderFieldEvent);
+
+        if (!$orderFieldEvent->value && $shippingMethod = $order->getShippingMethod()) {
+            $orderFieldEvent->value = $shippingMethod->handle;
         }
+
+        $this->addChildWithCDATA($order_xml, 'ShippingMethod', $orderFieldEvent->value);
     }
 
     /**
@@ -198,8 +226,8 @@ class Xml extends Component
 
         $item_xml->addChild('WeightUnits', $ss_weight_units);
 
-        if (isset($item->snapshot['options'])) {
-            $option_xml = $this->options($item_xml, $item->snapshot['options']);
+        if (isset($item->options)) {
+            $option_xml = $this->options($item_xml, $item->options);
         }
 
         return $item_xml;
@@ -302,7 +330,16 @@ class Xml extends Component
      */
     public function billTo(\SimpleXMLElement $customer_xml, Order $order, Customer $customer)
     {
-        if ($billingAddress = $order->getBillingAddress()) {
+        $billingAddress = $order->getBillingAddress();
+        if (!$billingAddress) {
+            $settings = Plugin::getInstance()->settings;
+            $billingSameAsShipping = $settings->billingSameAsShipping;
+            if ($billingSameAsShipping) {
+                $billingAddress = $order->getShippingAddress();
+            }
+        }
+
+        if ($billingAddress) {
             $billTo_xml = $this->address($customer_xml, $billingAddress, 'BillTo');
             if ($billingAddress->firstName && $billingAddress->lastName) {
                 $name = "{$billingAddress->firstName} {$billingAddress->lastName}";
@@ -385,25 +422,37 @@ class Xml extends Component
     public function customOrderFields(\SimpleXMLElement $order_xml, Order $order)
     {
         $customFields = [
-            'CustomField1',
-            'CustomField2',
-            'CustomField3',
-            'InternalNotes',
-            'CustomerNotes',
-            'Gift',
-            'GiftMessage'
+            OrderFieldEvent::FIELD_CUSTOM_FIELD_1,
+            OrderFieldEvent::FIELD_CUSTOM_FIELD_2,
+            OrderFieldEvent::FIELD_CUSTOM_FIELD_3,
+            OrderFieldEvent::FIELD_INTERNAL_NOTES,
+            OrderFieldEvent::FIELD_CUSTOMER_NOTES,
+            OrderFieldEvent::FIELD_GIFT,
+            OrderFieldEvent::FIELD_GIFT_MESSAGE,
         ];
-        // TODO:
-        // foreach ($customFields as $fieldName) {
-        //     if ($customFieldCallbacks = craft()->plugins->call("shipStationConnect{$fieldName}")) {
-        //         foreach ($customFieldCallbacks as $callback) {
-        //             if (is_callable($callback)) {
-        //                 $value = $callback($order);
-        //                 $order_xml->addChild($fieldName, substr(htmlspecialchars($value), 0, 100));
-        //             }
-        //         }
-        //     }
-        // }
+
+        foreach ($customFields as $fieldName) {
+            $orderFieldEvent = new OrderFieldEvent([
+                'field' => $fieldName,
+                'order' => $order,
+            ]);
+
+            Event::trigger(static::class, self::ORDER_FIELD_EVENT, $orderFieldEvent);
+            $data = $orderFieldEvent->value ?: '';
+
+            // Gift field requires a boolean value
+            if ($fieldName === OrderFieldEvent::FIELD_GIFT) {
+                $data = $data ? 'true' : 'false';
+                $order_xml->addChild($fieldName, $data);
+            } else {
+                if ($orderFieldEvent->cdata) {
+                    $this->addChildWithCDATA($order_xml, $fieldName, substr(htmlspecialchars($data), 0, 100));
+                } else {
+                    $order_xml->addChild($fieldName, substr(htmlspecialchars($data), 0, 100));
+                }
+            }
+        }
+
         return $order_xml;
     }